A2A & MCP/scenario1_local_agents/requirements.txt
# Scenario 1 Specific Requirements
# Inherits from main requirements.txt

# Azure AI and OpenAI
azure-ai-projects>=1.0.0
openai>=1.12.0
azure-identity>=1.12.0

# MCP SDK
mcp[cli]>=1.2.0

# HTTP and async
httpx>=0.25.0
aiohttp>=3.9.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# CLI and formatting
rich>=13.7.0
